backend/requirements.txt
# LLM & Agent Framework
google-generativeai>=0.3.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=0.0.5

# Data Processing
pandas>=2.0.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL
pymysql>=1.1.0          # MySQL
sshtunnel>=0.4.0        # SSH tunneling for remote DB access

# MCP Protocol
mcp>=1.23.0  # Required for November 2025 protocol update

# Async Database Drivers
asyncpg>=0.29.0  # Async PostgreSQL driver
aiomysql>=0.2.0  # Async MySQL driver

# Cloud Providers (for BigQuery, Snowflake, GCP Manager)
google-cloud-bigquery>=3.0.0
google-cloud-monitoring>=2.0.0
google-cloud-billing>=1.0.0
snowflake-connector-python>=3.6.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
pydantic>=2.0.0
rich>=13.0.0
textual>=0.40.0

# Web Framework (선택)
fastapi>=0.109.0
uvicorn>=0.27.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.8.0
